+++ app/Models/CocktailPrice.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kami\Cocktail\Models;
+
+readonly class CocktailPrice
+{
+    public function __construct(
+        public PriceCategory $priceCategory,
+        public Cocktail $cocktail,
+    ) {
+    }
+}